Rewrote the functions that return rating images to instead return an ordered list that can be styled with CSS
The `get_ratings_images()` and `get_ratings_images_vote()` functions
are rewritten to return an ordered list that can be styled with CSS
instead of images that can't. This makes it much easier for theme
authors to customize the look of the rate widget.

An added benefit is that the functions are now much shorter and with
less logic, which makes them easier to read and maintain.