Allow use inside custom Schema code
I wanted to use the plugin in a website where I already have schema code, but the plugin was duplicating the schema object, so I set the $itemtype = ' ' with wp_postratings_schema_itemtype filter 
Now I had 1 Schema object, but with duplicates values for title, link, description
so I update the code with a minuscule condition: if $itemtype is not set, don't even echo the meta. (because you already had it defined in yout template).